--- library/std/src/sys/pal/unix/linux/mod.rs
@@ -1,1 +1,0 @@
-pub mod pidfd;
