+++ library/std/src/sys/pal/unix/mod.rs
@@ -6,8 +6,6 @@ pub mod conf;
 #[cfg(target_os = "fuchsia")]
 pub mod fuchsia;
 pub mod futex;
-#[cfg(target_os = "linux")]
-pub mod linux;
 pub mod stack_overflow;
 pub mod sync;
 pub mod thread_parking;
